requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
discord==2.1.0
discord-py-interactions==4.4.0
discord.py==2.1.0
dnspython==2.3.0
Flask==2.2.3
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.3.3
pyparsing==3.0.9
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
